requirements.txt
pandas
numpy
matplotlib
scikit-learn
umap-learn
hdbscan
keybert
sentence-transformers
requests
python-dotenv
seaborn
streamlit
nltk
praw
